bumped up the number of adc samples per conversion cycle, dropped my phase A noise by a significant amount. Am observing a current magnitude difference; this may be due to layout. This is a concerning issue; I may want to spin a new revision with uniform shunt layout and positioning to see if there is some improvement to be had.

+++ Inc/foc_commutation.h
@@ -15,7 +15,7 @@ float gl_T;
 float gl_V;
 
 
-int gl_current_input_offset;
+int16_t gl_current_input_offset;
 float L;
 float R;
